+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - style pinned chats in chat list
 - `/pin` and `/unpin` commands to pin/unpin chats
 - `/mute` and `/unmute` commands to mute/unmute chats
+- `/topic` command to change chat name
 
 ## [0.5.0]